+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 *   Write JSON to the logger, rather than Ruby [Issue #430]
 *   Added ability to access all of a nodes attributes (Jon Rowe)
 *   Capybara 2.3 window support (Dmitry Vorotilin)
+*   Added ability to manipulate modals (`window.alert()`, `window.confirm()`
+    and `window.prompt()`) that has been implemented from Capybara 2.4
+    (Wataru MIYAGUNI)
 
 #### Bug fixes ####